Using libnss directly to recover the passwords
This is a big struggle and still one cannot decrypt password protected
databases. It's mainly inspired by a tool called "ffpwdump".
The calling options are a bit weird, but it will print an explanatory
text the way it is right now. One could certainly refactor that into the
program description and such a change would be very welcome.

It works well with unprotected databases and is very fast compared to
the other solution which needs to call out a lot. However, password
protected databases won't work because I don't know the proper API and a
quick look into sdrtest didn't help me. Mainly because they use SECU_*
functions which the libnss3.so on my system doesn't have.