Made it work with password protected directories
Finally, I found how to do it. It took quite a while. To my surprise
PK11_CheckUserPassword would also authenticate for the future. So we use
that function to not only test whether we need a password, but also to
open the encrypted container.

I tried various other methods like ChangeUserPassword or
SetPasswordFunc, but none of them would work. Plus it was a big hassle
using ctypes correctly. I am glad that it finally worked out so easily.